+++ packages/react-schema-form/src/Field/index.tsx
@@ -1,0 +1,1 @@
+export { default as createField } from './createField';
